railbird/requirements.txt
mss>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.16.0
eval7>=0.1.9
PyQt6>=6.5.0
pynput>=1.7.0
pygetwindow>=0.0.9
